Warnings cleaning

+++ include/array.hpp
@@ -1,13 +1,13 @@
 #pragma once
 
-template<typename T, uint32_t N>
+template<typename T, uint64_t N>
 struct Array
 {
-	T& operator[](const uint32_t index) {
+	T& operator[](const uint64_t index) {
 		return data[index];
 	}
 
-	T operator[](const uint32_t index) const {
+	T operator[](const uint64_t index) const {
 		return data[index];
 	}
